+++ src/main/kotlin/com/group4/ticketingservice/repository/BookmarkRepository.kt
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface BookmarkRepository : JpaRepository<Bookmark, Long> {
-    fun findByIdAndUserId(id: Int, userId: Long): Bookmark
-    fun deleteByIdAndUserId(id: Int, userId: Long)
-    fun findByUserId(userId: Long): List<Bookmark>
+interface BookmarkRepository : JpaRepository<Bookmark, Int> {
+    fun findByIdAndUserId(id: Int, userId: Int): Bookmark
+    fun deleteByIdAndUserId(id: Int, userId: Int)
+    fun findByUserId(userId: Int): List<Bookmark>
 }
