+++ src/main/kotlin/com/group4/ticketingservice/service/ReservationService.kt
@@ -19,7 +19,7 @@ class ReservationService @Autowired constructor(
     private val clock: Clock
 ) {
     @Transactional
-    fun createReservation(eventId: Long, userId: Long): Reservation {
+    fun createReservation(eventId: Int, userId: Int): Reservation {
         val user = userRepository.getReferenceById(userId)
         val event = eventRepository.findByIdWithPesimisticLock(eventId) ?: throw RuntimeException("")
 
@@ -36,13 +36,13 @@ class ReservationService @Autowired constructor(
         return reservation
     }
 
-    fun getReservation(reservationId: Long): Reservation {
+    fun getReservation(reservationId: Int): Reservation {
         return reservationRepository.findById(reservationId).orElseThrow {
             IllegalArgumentException("Reservation not found")
         }
     }
 
-    fun updateReservation(reservationId: Long, eventId: Long): Reservation {
+    fun updateReservation(reservationId: Int, eventId: Int): Reservation {
         val reservation: Reservation = reservationRepository.findById(reservationId).orElseThrow {
             IllegalArgumentException("Reservation not found")
         }
@@ -54,7 +54,7 @@ class ReservationService @Autowired constructor(
         return reservationRepository.save(reservation)
     }
 
-    fun deleteReservation(userId: Long, id: Long) {
+    fun deleteReservation(userId: Int, id: Int) {
         val reservation = reservationRepository.findByIdOrNull(id) ?: throw IllegalArgumentException("Reservation not found")
         if (reservation.user.id != userId) throw IllegalArgumentException("It's not your reservation")
         reservationRepository.delete(reservation)
